Implemented Dias's comments.

+++ backend/src/main/resources/db/migration/V4__add_card_table.sql
@@ -2,9 +2,9 @@ CREATE TABLE IF NOT EXISTS card(
     cardNumber INTEGER,
     ExpirationDate DATE,
     cvv INTEGER,
-    creditLimit INTEGER,
-    cardBalance INTEGER,
-    minimumPayment INTEGER,
-    rewardPoints INTEGER,
+    creditLimit NUMERIC(12, 2),
+    cardBalance NUMERIC(12, 2),
+    minimumPayment NUMERIC(8, 2),
+    rewardPoints NUMERIC(10, 2),
     customer_id INTEGER
 );
